+++ h3/sql/updates/h3--4.0.3--4.1.0.sql
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Bytes & Brains
+ * Copyright 2022-2023 Bytes & Brains
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
-\echo Use "ALTER EXTENSION h3 UPDATE TO 'unreleased'" to load this file. \quit
+\echo Use "ALTER EXTENSION h3 UPDATE TO '4.1.0'" to load this file. \quit
 
 DROP FUNCTION IF EXISTS h3_cell_to_boundary_wkb(h3index);
 DROP FUNCTION IF EXISTS h3_cells_to_multi_polygon_wkb(h3index[]);