+++ h3_postgis/sql/install/05-regions.sql
@@ -43,19 +43,19 @@ CREATE OR REPLACE FUNCTION h3_polygon_to_cells(multi geometry, resolution intege
 CREATE OR REPLACE FUNCTION h3_polygon_to_cells(multi geography, resolution integer) RETURNS SETOF h3index
 AS $$ SELECT h3_polygon_to_cells($1::geometry, $2) $$ LANGUAGE SQL IMMUTABLE PARALLEL SAFE CALLED ON NULL INPUT; -- NOT STRICT
 
---@ availability: unreleased
+--@ availability: 4.1.0
 --@ refid: h3_cells_to_multi_polygon_geometry
 CREATE OR REPLACE FUNCTION
     h3_cells_to_multi_polygon_geometry(h3index[]) RETURNS geometry
 AS $$ SELECT h3_cells_to_multi_polygon_wkb($1)::geometry $$ IMMUTABLE STRICT PARALLEL SAFE LANGUAGE SQL;
 
---@ availability: unreleased
+--@ availability: 4.1.0
 --@ refid: h3_cells_to_multi_polygon_geography
 CREATE OR REPLACE FUNCTION
     h3_cells_to_multi_polygon_geography(h3index[]) RETURNS geography
 AS $$ SELECT h3_cells_to_multi_polygon_wkb($1)::geography $$ IMMUTABLE STRICT PARALLEL SAFE LANGUAGE SQL;
 
---@ availability: unreleased
+--@ availability: 4.1.0
 --@ refid: h3_cells_to_multi_polygon_geometry_agg
 CREATE AGGREGATE h3_cells_to_multi_polygon_geometry(h3index) (
     sfunc = array_append,
@@ -64,7 +64,7 @@ CREATE AGGREGATE h3_cells_to_multi_polygon_geometry(h3index) (
     parallel = safe
 );
 
---@ availability: unreleased
+--@ availability: 4.1.0
 --@ refid: h3_cells_to_multi_polygon_geography_agg
 CREATE AGGREGATE h3_cells_to_multi_polygon_geography(h3index) (
     sfunc = array_append,